news_ai/requirements.txt
fastapi
uvicorn
transformers
torch
newspaper3k
deep-translator
cachetools
lxml